Resolves 50 conflicts (184 hunks) across 13 local commits and 20 upstream.
Bundle work stays ours; upstream wins where it is a fix or new infrastructure.

Absorbed from upstream: spaces (space_id), sharing grants, trash (soft
delete), templates, onboarding, authz, the fieldEditability field-metadata
API, and a squashed migration baseline (0001-0005).

Kept ours: PT linkage and artifact properties, dependency staleness
(built_revision / dependencies_current), Overture vocabularies and bundle
styling, member_draw_rank stacking, make_room_at_top layer ordering, the
bundle_type table removal, the edit-session fix and the PT filter gate.

Upstream preferred inside bundle code where its version is better:
- bundle_create_filtered refuses a filter that fails to parse, instead of
  silently copying the whole bundle and calling it filtered
- publish_artifact_if_current reads the displaced path under FOR UPDATE, so
  two builds of one revision cannot leak the first winner's archive
- an ambiguous import dispatch failure no longer deletes a bundle whose job
  may be running
- BundleFilter reuses the layer filter's expression helpers

Three bugs of ours that upstream's tests caught:
- build_mixin read layers_revision after the build, so output overtaken by a
  concurrent edit claimed the newer revision and passed the publish guard
- writer declared length_m as float32 while the file holds float64
- BundleRead.user_id was required, so one offboarded creator 422'd a listing

Four couplings no conflict marker showed: artifacts/__init__ imported a
helper base.py does not define; 12 authz test modules imported the deleted
seed_bundle_types; upstream's spy builders rejected the dependencies kwarg;
the BundleSummary fixture never set artifacts_from_layers.

Migrations squashed onto upstream's baseline: five local revisions become
0006_bundle_provenance (built_revision, bundle_artifact.properties, and the
bundle_type table drop). Two were redundant — upstream's 0001 already does
the status -> build_status rename and the 'processing' default. Verified on a
throwaway database: fresh chain, downgrade round-trip, prod-shape -> 0006,
and a re-run against an already-applied schema.

Verified: core 582 passed, geoapi 226 passed (9 await customer.space
locally), goatlib 1156 passed, web tsc/eslint clean, 1879 web tests passed.

+++ packages/python/goatlib/src/goatlib/bundles/importers/street_network/overture/writer.py
@@ -34,7 +34,16 @@ EDGE_SCHEMA = pa.schema(
         ("name", pa.string()),
         ("class", pa.string()),
         ("subclass", pa.string()),
-        ("length_m", pa.float32()),
+        # float64, not float32: the value is written by the `length` computed
+        # kind's own SQL, which is DOUBLE, and `SELECT * REPLACE (<expr> AS
+        # length_m)` takes the expression's type — so a narrower declaration
+        # here would simply be false about the file. Widening the declaration
+        # rather than narrowing the expression, because DOUBLE is also what the
+        # routing artifact reads the column as and what a later recompute
+        # produces: one width the whole way through, and no rounding step that
+        # would make the length a user sees differ from the length the engine
+        # routes on.
+        ("length_m", pa.float64()),
         ("surface", pa.string()),
         ("speed_limit_kph_forward", pa.int32()),
         ("speed_limit_kph_backward", pa.int32()),
